Hello, I've been entering data to create the database in sqlite for a long time, I tried to connect to this sqlite, I don't know if I succeeded. I am getting an error as follows Value must be ≥ 0 but getColumnIndex can be -1. I will try to fix this error tomorrow
Hello, I've been entering data to create the database in sqlite for a long time, I tried to connect to this sqlite, I don't know if I succeeded. I am getting an error as follows
Value must be ≥ 0 but getColumnIndex can be -1. I will try to fix this error tomorrow