I encountered a number of errors, it was an error about the theme file inflating when opening. After solving that error, I was able to take a screenshot on the screen. Even though it was a problem with the theme, it's still a fixable bug, I'll look at it tomorrow
I encountered a number of errors, it was an error about the theme file inflating when opening. After solving that error, I was able to take a screenshot on the screen. Even though it was a problem with the theme, it's still a fixable bug, I'll look at it tomorrow